refactor: split source layout per refactoring plan (phases 1-3, 7)
- Phase 1: remove mod declarations from main.rs; all modules now
  imported via rustinel:: from lib.rs as the single library root
- Phase 2: expand src/memory.rs into src/memory/ with platform
  sub-modules (linux, windows, unsupported) and a types module
- Phase 3 (partial): extract logging helpers into src/runtime/logging.rs
- Phase 7: update docs/roadmap.md and docs/limitations.md to reflect
  that YARA memory scanning is implemented, not planned

All tests pass on Windows (MSVC) and Linux (eBPF).
